Add initial CSV output mode.

+++ jpeginfo.c
@@ -46,7 +46,7 @@
 #define VERSION     "1.7.0beta"
 #define COPYRIGHT   "Copyright (C) 1996-2023 Timo Kokkonen"
 
-#define BUF_LINES   255
+#define BUF_LINES   512
 
 #ifndef HOST_TYPE
 #define HOST_TYPE ""
@@ -57,10 +57,6 @@
 #define METHODDEF(x) static x
 #endif
 
-#define EXIF_JPEG_MARKER   JPEG_APP0+1
-#define EXIF_IDENT_STRING  "Exif\000\000"
-#define EXIF_IDENT_STRING_LEN 6
-
 struct my_error_mgr {
 	struct jpeg_error_mgr pub;
 	jmp_buf setjmp_buffer;
@@ -69,6 +65,21 @@ typedef struct my_error_mgr * my_error_ptr;
 
 static struct jpeg_decompress_struct cinfo;
 static struct my_error_mgr jerr;
+
+struct jpeg_info {
+	int width;
+	int height;
+	int color_depth;
+	int progressive;
+	int check;
+	size_t size;
+	char *filename;
+	char *type;;
+	char *info;
+	char *comments;
+	char *digest;
+	char *error;
+};
 
 FILE *infile=NULL;
 FILE *listfile=NULL;
@@ -87,6 +98,7 @@ int input_from_file = 0;
 int md5_mode = 0;
 int sha256_mode = 0;
 int stdin_mode = 0;
+int csv_mode = 0;
 char *current = NULL;
 char last_error[JMSG_LENGTH_MAX + 1];
 
@@ -104,6 +116,7 @@ static struct option long_options[] = {
 	{"sha256",0,0,'2'},
 	{"version",0,0,'V'},
 	{"comments",0,0,'C'},
+	{"csv",0,0,'s'},
 	{"stdin",0,&stdin_mode,1},
 	{0,0,0,0}
 };
@@ -132,7 +145,7 @@ METHODDEF(void)
 
 	global_error_counter++;
 	global_total_errors++;
-	strncpy(last_error, buffer, sizeof(last_error));
+	strncopy(last_error, buffer, sizeof(last_error));
 }
 
 
@@ -177,28 +190,29 @@ void print_usage(void)
 	fprintf(stderr,"jpeginfo v" VERSION " " COPYRIGHT "\n");
 	fprintf(stderr,
 		"Usage: jpeginfo [options] <filenames>\n\n"
-		"  -2, --sha256    calculate SHA-256 checksum for each file\n"
-		"  -5, --md5       calculate MD5 checksum for each file\n"
-		"  -c, --check     check files also for errors\n"
-		"  -C, --comments  display comments (from COM markers)\n"
-		"  -d, --delete    delete files that have errors\n"
+		"  -2, --sha256    Calculate SHA-256 checksum for each file.\n"
+		"  -5, --md5       Calculate MD5 checksum for each file.\n"
+		"  -c, --check     Check files also for errors.\n"
+		"  -C, --comments  Display comments (from COM markers)\n"
+		"  -d, --delete    Delete files that have errors\n"
 		"  -f<filename>,  --file<filename>\n"
-		"                  read the filenames to process from given file\n"
+		"                  Read the filenames to process from given file\n"
 		"                  (for standard input use '-' as a filename)\n"
-		"  -h, --help      display this help and exit\n"
-		"  -i, --info      display even more information about pictures\n"
-		"  -l, --lsstyle   use alternate listing format (ls -l style)\n"
+		"  -h, --help      Display this help and exit\n"
+		"  -i, --info      Display even more information about pictures\n"
+		"  -l, --lsstyle   Use alternate listing format (ls -l style)\n"
 		"  -m<mode>, --mode=<mode>\n"
-		"                  defines which jpegs to remove (when using"
+		"                  Defines which jpegs to remove (when using"
 		" the -d option).\n"
 		"                  Mode can be one of the following:\n"
 		"                    erronly     only files with serious errrors\n"
 		"                    all         files ontaining warnings or errors (default)\n"
-		"  -q, --quiet     quiet mode, output just jpeg infos\n"
-		"  -v, --verbose   enable verbose mode (positively chatty)\n"
-		"  -V, --version	  print program version and exit\n"
+		"  -q, --quiet     Quiet mode, output just jpeg infos\n"
+		"  -s, --csv       Comma separated (CSV) output style.\n"
+		"  -v, --verbose   Enable verbose mode (positively chatty)\n"
+		"  -V, --version	  Print program version and exit\n"
 		"\n"
-		"   -, --stdin         read input from standard input (instead of a file)\n"
+		"   -, --stdin     Read input from standard input (instead of a file)\n"
 		"\n\n");
 
 	exit(0);
@@ -217,7 +231,7 @@ void parse_args(int argc, char **argv)
 
 	while(1) {
 		opt_index=0;
-		if ( (c=getopt_long(argc,argv, "livVdcChqm:f:52",
+		if ( (c=getopt_long(argc,argv, "livVdcChqm:f:52s",
 						long_options, &opt_index))  == -1)
 			break;
 		switch (c) {
@@ -268,6 +282,9 @@ void parse_args(int argc, char **argv)
 		case 'C':
 			com_mode=1;
 			break;
+		case 's':
+			csv_mode=1;
+			break;
 		case '?':
 			break;
 
@@ -288,44 +305,248 @@ void parse_args(int argc, char **argv)
 }
 
 
+void clear_jpeg_info(struct jpeg_info *info)
+{
+	if (!info)
+		return;
+
+	memset(info, 0, sizeof(struct jpeg_info));
+}
+
+
+void free_jpeg_info(struct jpeg_info *info)
+{
+	if (!info)
+		return;
+
+	if (info->filename)
+		free(info->filename);
+	if (info->type)
+		free(info->type);
+	if (info->info)
+		free(info->info);
+	if (info->comments)
+		free(info->comments);
+	if (info->digest)
+		free(info->digest);
+	if (info->error)
+		free(info->error);
+
+	clear_jpeg_info(info);
+}
+
+
+void parse_jpeg_info(struct jpeg_decompress_struct *cinfo, struct jpeg_info *info)
+{
+	jpeg_saved_marker_ptr cmarker;
+	int marker_in_count, marker_in_size;
+	char marker_str[256], comment_str[1024], tmp[64];
+
+	if (!cinfo || !info)
+		return;
+
+	info->width = (int)cinfo->image_width;
+	info->height = (int)cinfo->image_height;
+	info->color_depth = (int)cinfo->num_components * 8;
+	info->progressive = (cinfo->progressive_mode ? 1 : 0);
+
+	if (cinfo->saw_Adobe_marker)
+		info->type = strdup("Adobe");
+	else if (cinfo->saw_JFIF_marker)
+		info->type = strdup("JFIF");
+	else
+		info->type = strdup("n/a");
+
+	strncopy(marker_str, (cinfo->arith_code ? "Arithmetic" : "Huffman"), sizeof(marker_str));
+	comment_str[0]=0;
+
+	/* Check for Exif/IPTC/ICC/XMP/etc. markers */
+	marker_in_count=0;
+	marker_in_size=0;
+	cmarker=cinfo->marker_list;
+
+	while (cmarker) {
+		marker_in_count++;
+		marker_in_size+=cmarker->data_length;
+
+		if (cmarker->marker == EXIF_JPEG_MARKER &&
+			cmarker->data_length >= EXIF_IDENT_STRING_SIZE &&
+			!memcmp(cmarker->data,EXIF_IDENT_STRING,EXIF_IDENT_STRING_SIZE))
+			str_add_list(marker_str, sizeof(marker_str), "Exif", ",");
+
+		if (cmarker->marker == IPTC_JPEG_MARKER)
+			str_add_list(marker_str, sizeof(marker_str), "IPTC", ",");
+
+		if (cmarker->marker == ICC_JPEG_MARKER &&
+			cmarker->data_length >= ICC_IDENT_STRING_SIZE &&
+			!memcmp(cmarker->data,ICC_IDENT_STRING,ICC_IDENT_STRING_SIZE))
+			str_add_list(marker_str, sizeof(marker_str), "ICC", ",");
+
+		if (cmarker->marker == XMP_JPEG_MARKER &&
+			cmarker->data_length >= XMP_IDENT_STRING_SIZE &&
+			!memcmp(cmarker->data,XMP_IDENT_STRING,XMP_IDENT_STRING_SIZE))
+			str_add_list(marker_str, sizeof(marker_str), "XMP", ",");
+
+		if (cmarker->marker == JFXX_JPEG_MARKER &&
+			cmarker->data_length >= JFXX_IDENT_STRING_SIZE &&
+			!memcmp(cmarker->data, JFXX_IDENT_STRING, JFXX_IDENT_STRING_SIZE))
+			str_add_list(marker_str, sizeof(marker_str), "JFXX", ",");
+
+		if (cmarker->marker == JPEG_COM && cmarker->data_length > 0) {
+			int o = 0;
+			for (int i = 0; i < cmarker->data_length; i++) {
+				char ch = cmarker->data[i];
+				if (ch < 32 || iscntrl(ch)) {
+					ch = '.';
+				}
+				if (o < sizeof(tmp) - 1) {
+					*(tmp + o++) = ch;
+				}
+			}
+			*(tmp + o) = 0;
+
+			str_add_list(comment_str, sizeof(comment_str), tmp, ",");
+		}
+
+		cmarker=cmarker->next;
+	}
+
+	if (cinfo->density_unit == 1 || cinfo->density_unit == 2) {
+		snprintf(tmp, sizeof(tmp), "%ddp%c", MIN(cinfo->X_density, cinfo->Y_density),
+			(cinfo->density_unit == 1 ? 'i' : 'c') );
+		str_add_list(marker_str, sizeof(marker_str), tmp, ",");
+	}
+
+	if (cinfo->CCIR601_sampling) {
+		str_add_list(marker_str, sizeof(marker_str), "CCIR601", ",");
+	}
+
+	info->info = strdup(marker_str);
+	info->comments = strdup(comment_str);
+}
+
+const char *check_status_str(int check)
+{
+	if (check == 1)
+		return "OK";
+	else if (check == 2)
+		return "WARNING";
+	else if (check == 3)
+		return "ERROR";
+
+	return "";
+}
+
+void print_jpeg_info(struct jpeg_info *info)
+{
+	const char *type, *einfo, *com, *error, *digest;
+	char p;
+
+	if (!info)
+		return;
+
+	if (quiet_mode > 1)
+		return;
+
+	type = (info->type ? info->type : "");
+	einfo = (info->info ? info->info : "");
+	com = (info->comments ? info->comments : "");
+	error = (info->error ? info->error : "");
+	digest = (info->digest ? info->digest : "");
+
+	p = (info->progressive ? 'P' : 'N');
+
+	if (csv_mode) {
+		printf("\"%s\",%lu,\"%s\",%d,%d,\"%dbit\",\"%s\",\"%c\",\"%s\",\"%s\",\"%s\",\"%s\"\n",
+			info->filename,
+			info->size,
+			digest,
+			info->width,
+			info->height,
+			info->color_depth,
+			type,
+			p,
+			einfo,
+			com,
+			check_status_str(info->check),
+			error
+			);
+	}
+	else if (list_mode) {
+		printf("%4d x %4d %2dbit %5s %c%s%s %7lu %s%s%s %s%s%s\n",
+			info->width,
+			info->height,
+			info->color_depth,
+			type,
+			p,
+			(longinfo_mode ? " " : ""),
+			(longinfo_mode ? einfo : ""),
+			info->size,
+			digest,
+			(info->digest ? " " : ""),
+			info->filename,
+			check_status_str(info->check),
+			(info->error ? " " : ""),
+			error
+			);
+	}
+	else {
+		printf("%s %4d x %4d %2dbit %5s %c%s%s %7lu %s%s%s%s%s\n",
+			info->filename,
+			info->width,
+			info->height,
+			info->color_depth,
+			type,
+			p,
+			(longinfo_mode ? " " : ""),
+			(longinfo_mode ? einfo : ""),
+			info->size,
+			digest,
+			(info->digest ? " " : ""),
+			check_status_str(info->check),
+			(info->error ? " " : ""),
+			error
+			);
+	}
+}
+
 /*****************************************************************************/
 int main(int argc, char **argv)
 {
 	MD5_CTX *MD5 = malloc(sizeof(MD5_CTX));
 	JSAMPARRAY buf = malloc(sizeof(JSAMPROW)*BUF_LINES);
-	jpeg_saved_marker_ptr exif_marker, cmarker;
 	volatile int i;
 	int j;
-	unsigned char ch;
 	char namebuf[1024];
 	unsigned char *inbuf = NULL;
 	long long fs;
 	char digest_text[65];
 	size_t inbuffersize;
-	volatile int phase;
+	struct jpeg_info *info = malloc(sizeof(struct jpeg_info));
 
-	if (!buf || !MD5) no_memory();
+	if (!buf || !MD5 || !info)
+		no_memory();
 
 	for(i = 0; i < BUF_LINES; i++) {
 		buf[i] = NULL;
 	}
+	clear_jpeg_info(info);
 
 	cinfo.err = jpeg_std_error(&jerr.pub);
 	jpeg_create_decompress(&cinfo);
 	jerr.pub.error_exit=my_error_exit;
 	jerr.pub.output_message=my_output_message;
 
-	/* parse command line parameters */
+	/* Parse command line parameters */
 	parse_args(argc, argv);
 	i=(optind > 0 ? optind : 1);
 
-	/* process input file(s) */
+	/* Loop to process input file(s) */
 	do {
-		phase = 0;
+		free_jpeg_info(info);
 		current = argv[i];
 
-		/* open input file */
-
+		/* Open input file */
 		if (stdin_mode) {
 			if (verbose_mode)
 				fprintf(stderr, "Reading file: <STDIN>\n");
@@ -348,21 +569,27 @@ int main(int argc, char **argv)
 			}
 			if (is_dir(infile)) {
 				fclose(infile);
-				if (verbose_mode) printf("directory: %s  skipped\n", current);
+				if (verbose_mode) fprintf(stderr, "Skipping directory: %s\n", current);
 				continue;
 			}
 			inbuffersize = filesize(infile);
 		}
+		info->filename = strdup(current);
 
-		/* read input file to buffer */
+		/* Read input file into a buffer */
 		fs = read_file(infile, (inbuffersize > 0 ? inbuffersize : 65536), &inbuf);
 		if (fs < 0) {
 			no_memory();
 			continue;
 		}
 		fclose(infile);
+		info->size = fs;
 
+		/* Error handler for (libjpeg) errors in decoding */
+		last_error[0] = 0;
 		if (setjmp(jerr.setjmp_buffer)) {
+			info->check = 3;
+			info->error = strdup(last_error);
 			jpeg_abort_decompress(&cinfo);
 			for(j = 0; j < BUF_LINES; j++) {
 				if (buf[j]) {
@@ -371,24 +598,14 @@ int main(int argc, char **argv)
 				}
 			}
 			if (quiet_mode < 2) {
-				if (phase < 1)
-					printf("%s", current);
-				else if (phase < 2) {
-					printf("   0 x 0    n/a   n/a   N %7d", fs);
-					if (md5_mode | sha256_mode)
-						printf(" %s", digest_text);
-					if (list_mode)
-						printf(" %s ", current);
-				}
-				printf(" [ERROR] %s\n", last_error);
+				print_jpeg_info(info);
 			}
 			if (delete_mode) delete_file(current, verbose_mode, quiet_mode);
 			continue;
 		}
 
-
+		/* Calculate hash (message-digest) of the input file */
 		if (md5_mode || sha256_mode) {
-			/* calculate hash (message-digest) of the input file */
 			unsigned char digest[32];
 
 			digest_text[0] = 0;
@@ -401,71 +618,20 @@ int main(int argc, char **argv)
 				crypto_hash_sha256(digest, inbuf, fs);
 				digest2str(digest, digest_text, 32);
 			}
+			info->digest = strdup(digest_text);
 		}
 
-		if (!list_mode && quiet_mode < 2) printf("%s ", current);
-		phase = 1;
 
+		/* Read JPEG file header */
 		global_error_counter=0;
 		if (com_mode) jpeg_save_markers(&cinfo, JPEG_COM, 0xffff);
 		jpeg_save_markers(&cinfo, EXIF_JPEG_MARKER, 0xffff);
 		jpeg_mem_src(&cinfo, inbuf, fs);
 		jpeg_read_header(&cinfo, TRUE);
+		parse_jpeg_info(&cinfo, info);
 
-		/* check for Exif marker */
-		exif_marker=NULL;
-		cmarker=cinfo.marker_list;
-		while (cmarker) {
-			if (cmarker->marker == EXIF_JPEG_MARKER && cmarker->data_length >= EXIF_IDENT_STRING_LEN) {
-				if (!memcmp(cmarker->data, EXIF_IDENT_STRING, EXIF_IDENT_STRING_LEN))
-					exif_marker=cmarker;
-			}
-			cmarker=cmarker->next;
-		}
 
-		if (quiet_mode < 2) {
-			printf("%4d x %-4d %2dbit ", (int)cinfo.image_width,
-				(int)cinfo.image_height, (int)cinfo.num_components*8);
-
-			if (exif_marker) printf("Exif  ");
-			else if (cinfo.saw_JFIF_marker) printf("JFIF  ");
-			else if (cinfo.saw_Adobe_marker) printf("Adobe ");
-			else printf("n/a   ");
-
-			if (longinfo_mode) {
-				printf("%s %s", (cinfo.progressive_mode?"Progressive":"Normal"),
-					(cinfo.arith_code?"Arithmetic":"Huffman") );
-
-				if (cinfo.density_unit==1||cinfo.density_unit==2)
-					printf(",%ddp%c", MIN(cinfo.X_density,cinfo.Y_density),
-						(cinfo.density_unit==1?'i':'c') );
-
-				if (cinfo.CCIR601_sampling) printf(",CCIR601");
-				printf(" %7lld ",fs);
-
-			} else printf("%c %7lld ",(cinfo.progressive_mode ? 'P' : 'N'), fs);
-
-			if (md5_mode || sha256_mode) printf("%s ", digest_text);
-			if (list_mode) printf("%s ", current);
-
-			if (com_mode) {
-				cmarker=cinfo.marker_list;
-				while (cmarker) {
-					if (cmarker->marker == JPEG_COM) {
-						printf("\"");
-						for (j=0; j < cmarker->data_length; j++) {
-							ch = cmarker->data[j];
-							if (ch < 32 || iscntrl(ch)) continue;
-							printf("%c", cmarker->data[j]);
-						}
-						printf("\" ");
-					}
-					cmarker=cmarker->next;
-				}
-			}
-		}
-		phase = 2;
-
+		/* Decode JPEG to check for errors in the file */
 		if (check_mode) {
 			cinfo.out_color_space = JCS_GRAYSCALE; /* to speed up the process... */
 			cinfo.scale_denom = 8;
@@ -489,23 +655,21 @@ int main(int argc, char **argv)
 				buf[j] = NULL;
 			}
 
-			if (!global_error_counter) {
-				if (quiet_mode < 2) printf(" [OK]\n");
-			}
-			else {
-				if (quiet_mode < 2) printf(" [WARNING]\n");
-				if (delete_mode && !del_mode)
+			info->check = (global_error_counter == 0 ? 1 : 2);
+			info->error = strdup(last_error);
+			print_jpeg_info(info);
+			if (delete_mode && !del_mode && info->check > 1)
 					delete_file(current, verbose_mode, quiet_mode);
-			}
 		}
 		else { /* !check_mode */
-			if (quiet_mode < 2) printf("\n");
+			print_jpeg_info(info);
 			jpeg_abort_decompress(&cinfo);
 		}
 
 
 	} while ((!stdin_mode && ++i<argc) || input_from_file);
 
+	free_jpeg_info(info);
 	if (inbuf)
 		free(inbuf);
 	jpeg_destroy_decompress(&cinfo);
